(refactor/src_file_names) remove empty req_handler.cpp and .hpp and rename http_request -> htpp_request_handler and http -> http_request
