+++ src/get_request.cpp
@@ -1,4 +1,4 @@
-#include "http_request.hpp"
+#include "http_request_handler.hpp"
 #include "error.hpp"
 
 GetRequestHandler::GetRequestHandler(const HttpRequest &m) : HttpRequestHandler(m) {}
