+++ src/server.cpp
@@ -1,7 +1,5 @@
-#include "req_handler.hpp"
 #include "server.hpp"
-#include "http.hpp"
-#include "http_request.hpp"
+#include "http_request_handler.hpp"
 #include "log.hpp"
 #include "socket.hpp"
 #include "utils.hpp"
